Match relations to bindings using network_get
The `ingress-address` sent over the relation data is the address from
the remote charm's perspective, which has no guarantee of actually being
on any subnet available to the local charm (only that it will be
routable). Instead, we have to use `network_get` to match the relation
to the space binding based on the egress-subnets.

Additionally, `network-get --primary-address` is deprecated, so we can't
use that, either.

Closes-Bug: 1843809